Expand targeting tests. Move prebid targeting into targeting file.
